refactor: 히스토리 조회 response에 해당 면담의 시트가 포함되도록 수정
